+++ errors/errgo.go
@@ -1,16 +1,22 @@
 package errors
 
-import "gopkg.in/errgo.v1"
+import (
+	"gopkg.in/errgo.v1"
+)
 
 func errgoRoot(err error) error {
 	for {
-		e, ok := err.(*errgo.Err)
+		e, ok := err.(ErrCtx)
+		if ok {
+			err = e.err
+		}
+		errgoErr, ok := err.(*errgo.Err)
 		if !ok {
 			return err
 		}
-		if e.Underlying() == nil {
+		if errgoErr.Underlying() == nil {
 			return err
 		}
-		err = e.Underlying()
+		err = errgoErr.Underlying()
 	}
 }