+++ errors/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## To be Released
 
+* feat: add New method to `ErrCtx`
+* feat: IsRootCause and RootCause are taking in account `ErrCtx` underlying errors
+
 ## v2.0.0
 
 * fix: privatify `ErrgoRoot`