+++ errors/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+## v2.1.0
+
 * feat: add New function to `ErrCtx`
 * feat: IsRootCause and RootCause are taking in account `ErrCtx` underlying errors
 * feat: RootCtxOrFallback retrieves the deepest context from wrapped errors.